refactor: review feedback
